Move error checking task to a subpackage, and create API in base class to also create Warnings
