+++ jsettlers.mapcreator/src/jsettlers/mapcreator/tools/landscape/FixHeightsTool.java
@@ -19,8 +19,8 @@ import jsettlers.common.position.ShortPoint2D;
 import jsettlers.mapcreator.data.LandscapeConstraint;
 import jsettlers.mapcreator.data.MapData;
 import jsettlers.mapcreator.data.objects.ObjectContainer;
-import jsettlers.mapcreator.mapvalidator.tasks.ValidateLandscape;
-import jsettlers.mapcreator.mapvalidator.tasks.ValidateResources;
+import jsettlers.mapcreator.mapvalidator.tasks.error.ValidateLandscape;
+import jsettlers.mapcreator.mapvalidator.tasks.error.ValidateResources;
 import jsettlers.mapcreator.tools.AbstractTool;
 import jsettlers.mapcreator.tools.shapes.ShapeType;
